refactor: move inputbox to widgets folder
